Major clean up and more documentation
I cleaned up my EnergyPlusFixture. I pulled the embedded classes out
into their own files. This allowed for better (and cleaner) testing.
