+++ tst/EnergyPlus/unit/DXCoils.unit.cc
@@ -24,7 +24,10 @@ using namespace OutputReportPredefined;
 
 namespace EnergyPlus {
 
-	TEST_F( HVACFixture, DXCoils_Test1 )
+	typedef HVACFixture DXCoilsFixture;
+	typedef HVACFixture DXCoilsDeathTestFixture;
+
+	TEST_F( DXCoilsFixture, Test1 )
 	{
 		using CurveManager::Quadratic;
 		using CurveManager::BiQuadratic;
@@ -159,7 +162,7 @@ namespace EnergyPlus {
 		PerfCurve.deallocate();
 
 	}
-	TEST_F( HVACFixture, DXCoils_Test2 )
+	TEST_F( DXCoilsFixture, Test2 )
 	{
 		using CurveManager::Quadratic;
 		using CurveManager::BiQuadratic;
