+++ tst/EnergyPlus/unit/IOHelper.unit.cc
@@ -4,16 +4,14 @@
 #include <gtest/gtest.h>
 
 // EnergyPlus Headers
-#include "Fixtures/IdfParserFixture.hh"
-#include "Fixtures/EnergyPlusMetaFixture.hh"
-
-// These must be here otherwise the linker will complain. This is so the shared resource m_idd_cache can be used by unit tests.
-struct InputProcessorCache;
-std::unique_ptr<EnergyPlus::EnergyPlusFixture::InputProcessorCache> EnergyPlus::EnergyPlusFixture::m_idd_cache = nullptr;
+#include "Fixtures/EnergyPlusFixture.hh"
 
 namespace EnergyPlus {
 
-	TEST_F( EnergyPlusMetaFixture, compareIDF )
+	typedef EnergyPlusFixture IOHelperFixture;
+	typedef EnergyPlusFixture IOHelperDeathTestFixture;
+
+	TEST_F( IOHelperFixture, compareIDF )
 	{
 		std::string const idf_objects = delimited_string({
 			"Version,8.3;",
@@ -55,14 +53,14 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( EnergyPlusMetaFixture, processIDD_Full_IDD )
+	TEST_F( IOHelperFixture, processIDD_Full_IDD )
 	{
 		using namespace InputProcessor;
 		std::string const idd_objects;
 
 		bool errors_found = false;
 
-		ASSERT_FALSE( process_idd( idd_objects, errors_found ) );
+		ASSERT_FALSE( IOHelper::process_idd( idd_objects, errors_found ) );
 
 		EXPECT_EQ( 745, NumObjectDefs );
 		ASSERT_EQ( static_cast<unsigned long>( NumObjectDefs ), ListOfObjects.size() );
@@ -95,7 +93,7 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( EnergyPlusMetaFixture, processIDD )
+	TEST_F( IOHelperFixture, processIDD )
 	{
 		using namespace InputProcessor;
 		std::string const idd_objects = delimited_string({
@@ -110,7 +108,7 @@ namespace EnergyPlus {
 
 		bool errors_found = false;
 
-		ASSERT_FALSE( process_idd( idd_objects, errors_found ) );
+		ASSERT_FALSE( IOHelper::process_idd( idd_objects, errors_found ) );
 
 		EXPECT_EQ( 1, NumObjectDefs );
 
@@ -137,7 +135,7 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( EnergyPlusMetaFixture, processIDF )
+	TEST_F( IOHelperFixture, processIDF )
 	{
 		using namespace InputProcessor;
 		std::string const idf_objects = delimited_string({
@@ -192,7 +190,7 @@ namespace EnergyPlus {
 
 	}
 
-	TEST_F( EnergyPlusMetaFixture, processIDF_Cached )
+	TEST_F( IOHelperFixture, processIDF_Cached )
 	{
 		using namespace InputProcessor;
 		std::string const idf_objects = delimited_string({
@@ -275,265 +273,6 @@ namespace EnergyPlus {
 		EXPECT_TRUE( compare_containers< std::vector< bool > >( { false, false, false, false, false, true, false, false }, IDFRecords( index ).AlphBlank ) );
 		EXPECT_TRUE( compare_containers< std::vector< Real64 > >( { 0.5, 4, 0, 0, 4.572, 0, 0, 0, 15.24, 0, 0, 15.24, 0, 4.572 }, IDFRecords( index ).Numbers ) );
 		EXPECT_TRUE( compare_containers< std::vector< bool > >( { false, false, false, false, false, false, false, false, false, false, false, false, false, false }, IDFRecords( index ).NumBlank ) );
-
-	}
-
-	TEST_F( IdfParserFixture, decode ) {
-		std::string const test_object( delimited_string( {
-			"Version,8.3;",
-			"  Building,",
-			"    Ref Bldg Medium Office New2004_v1.3_5.0,  !- Name",
-			"    0.0000,                  !- North Axis {deg}",
-			"    City,                    !- Terrain",
-			"    0.0400,                  !- Loads Convergence Tolerance Value",
-			"    0.2000,                  !- Temperature Convergence Tolerance Value {deltaC}",
-			"    FullInteriorAndExterior, !- Solar Distribution",
-			"    25,                      !- Maximum Number of Warmup Days",
-			"    6;  ",
-		} ) );
-
-		auto const output = decode( test_object );
-
-		EXPECT_EQ( std::vector< std::vector< std::string > >({ { "Version", "8.3" }, { "Building", "Ref Bldg Medium Office New2004_v1.3_5.0", "0.0000", "City", "0.0400", "0.2000", "FullInteriorAndExterior", "25", "6" } }), output );
-	}
-
-	TEST_F( IdfParserFixture, decode_success ) {
-		bool success = true;
-		std::string const test_object( delimited_string( {
-			"Version,8.3;",
-			"  Building,",
-			"    Ref Bldg Medium Office New2004_v1.3_5.0,  !- Name",
-			"    0.0000,                  !- North Axis {deg}",
-			"    City,                    !- Terrain",
-			"    0.0400,                  !- Loads Convergence Tolerance Value",
-			"    0.2000,                  !- Temperature Convergence Tolerance Value {deltaC}",
-			"    FullInteriorAndExterior, !- Solar Distribution",
-			"    25,                      !- Maximum Number of Warmup Days",
-			"    6;  ",
-		} ) );
-
-		auto const output = decode( test_object, success );
-
-		EXPECT_EQ( std::vector< std::vector< std::string > >({ { "Version", "8.3" }, { "Building", "Ref Bldg Medium Office New2004_v1.3_5.0", "0.0000", "City", "0.0400", "0.2000", "FullInteriorAndExterior", "25", "6" } }), output );
-		EXPECT_TRUE( success );
-	}
-
-	TEST_F( IdfParserFixture, parse_idf ) {
-		size_t index = 0;
-		bool success = true;
-		std::string const test_object( delimited_string( {
-			"Version,8.3;",
-			"  Building,",
-			"    Ref Bldg Medium Office New2004_v1.3_5.0,  !- Name",
-			"    0.0000,                  !- North Axis {deg}",
-			"    City,                    !- Terrain",
-			"    0.0400,                  !- Loads Convergence Tolerance Value",
-			"    0.2000,                  !- Temperature Convergence Tolerance Value {deltaC}",
-			"    FullInteriorAndExterior, !- Solar Distribution",
-			"    25,                      !- Maximum Number of Warmup Days",
-			"    6;  ",
-		} ) );
-
-		auto const output = parse_idf( test_object, index, success );
-
-		EXPECT_EQ( std::vector< std::vector< std::string > >({ { "Version", "8.3" }, { "Building", "Ref Bldg Medium Office New2004_v1.3_5.0", "0.0000", "City", "0.0400", "0.2000", "FullInteriorAndExterior", "25", "6" } }), output );
-		EXPECT_EQ( 434ul, index );
-		EXPECT_TRUE( success );
-	}
-
-	TEST_F( IdfParserFixture, parse_object ) {
-		size_t index = 0;
-		bool success = true;
-		std::string const test_object( delimited_string( {
-			"  Building,",
-			"    Ref Bldg Medium Office New2004_v1.3_5.0,  !- Name",
-			"    0.0000,                  !- North Axis {deg}",
-			"    City,                    !- Terrain",
-			"    0.0400,                  !- Loads Convergence Tolerance Value",
-			"    0.2000,                  !- Temperature Convergence Tolerance Value {deltaC}",
-			"    FullInteriorAndExterior, !- Solar Distribution",
-			"    25,                      !- Maximum Number of Warmup Days",
-			"    6;",
-		} ) );
-
-		auto const output_vector = parse_object( test_object, index, success );
-		EXPECT_EQ( std::vector< std::string >({ "Building", "Ref Bldg Medium Office New2004_v1.3_5.0", "0.0000", "City", "0.0400", "0.2000", "FullInteriorAndExterior", "25", "6" }), output_vector );
-		EXPECT_EQ( test_object.size() - 1, index );
-		EXPECT_TRUE( success );
-	}
-
-	TEST_F( IdfParserFixture, eat_whitespace ) {
-		size_t index = 0;
-		eat_whitespace( "    test", index );
-		EXPECT_EQ( 4ul, index );
-
-		index = 0;
-		eat_whitespace( "t   test", index );
-		EXPECT_EQ( 0ul, index );
-	}
-
-	TEST_F( IdfParserFixture, eat_comment ) {
-		size_t index = 0;
-		eat_comment( "!- North Axis {deg}\n", index );
-		EXPECT_EQ( 20ul, index );
-
-		index = 0;
-		eat_comment( "                    !- Terrain\n", index );
-		EXPECT_EQ( 31ul, index );
-
-		index = 0;
-		eat_comment( "  !- Name\n    0.0000", index );
-		EXPECT_EQ( 10ul, index );
-
-		index = 0;
-		eat_comment( "  !- Name\n\r    0.0000", index );
-		EXPECT_EQ( 10ul, index );
-	}
-
-	TEST_F( IdfParserFixture, parse_string ) {
-		size_t index = 0;
-		bool success = true;
-		std::string output_string;
-
-		output_string = parse_string( "test_string", index, success );
-		EXPECT_EQ( "test_string", output_string );
-		EXPECT_EQ( 11ul, index );
-		EXPECT_TRUE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_string( "test string", index, success );
-		EXPECT_EQ( "test string", output_string );
-		EXPECT_EQ( 11ul, index );
-		EXPECT_TRUE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_string( "-1234.1234", index, success );
-		EXPECT_EQ( "-1234.1234", output_string );
-		EXPECT_EQ( 10ul, index );
-		EXPECT_TRUE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_string( R"(\b\t/\\\";)", index, success );
-		EXPECT_EQ( "\b\t/\\\"", output_string );
-		EXPECT_EQ( 9ul, index );
-		EXPECT_TRUE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_string( R"(test \n string)", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 7ul, index );
-		EXPECT_FALSE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_string( R"(! this is a comment \n)", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 0ul, index );
-		EXPECT_TRUE( success );
-	}
-
-	TEST_F( IdfParserFixture, parse_value ) {
-		size_t index = 0;
-		bool success = true;
-		std::string output_string;
-
-		output_string = parse_value( "test_string", index, success );
-		EXPECT_EQ( "test_string", output_string );
-		EXPECT_EQ( 11ul, index );
-		EXPECT_TRUE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_value( ", test_string", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 0ul, index );
-		EXPECT_FALSE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_value( R"(test \n string)", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 7ul, index );
-		EXPECT_FALSE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_value( "; test_string", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 0ul, index );
-		EXPECT_FALSE( success );
-
-		index = 0;
-		success = true;
-		output_string = parse_value( "! test_string", index, success );
-		EXPECT_EQ( "", output_string );
-		EXPECT_EQ( 0ul, index );
-		EXPECT_FALSE( success );
-	}
-
-	TEST_F( IdfParserFixture, look_ahead ) {
-		std::string const test_input( "B , ! t ; `" );
-		size_t index = 0;
-		size_t token = look_ahead( test_input, index );
-		EXPECT_EQ( 0ul, index );
-		EXPECT_EQ( 5ul, token );
-		index = 2;
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( 2ul, index );
-		EXPECT_EQ( 3ul, token );
-		index = 3;
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( 3ul, index );
-		EXPECT_EQ( 2ul, token );
-		index = 5;
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( 5ul, index );
-		EXPECT_EQ( 5ul, token );
-		index = 7;
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( 7ul, index );
-		EXPECT_EQ( 4ul, token );
-		index = 9;
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( 9ul, index );
-		EXPECT_EQ( 0ul, token );
-		index = test_input.size();
-		token = look_ahead( test_input, index );
-		EXPECT_EQ( test_input.size(), index );
-		EXPECT_EQ( 1ul, token );
-
-	}
-
-	TEST_F( IdfParserFixture, next_token ) {
-		size_t index = 0;
-
-		std::string const test_input( "B , ! t ; `" );
-		size_t token = next_token( test_input, index );
-		EXPECT_EQ( 1ul, index );
-		EXPECT_EQ( 5ul, token );
-		token = next_token( test_input, index );
-		EXPECT_EQ( 3ul, index );
-		EXPECT_EQ( 3ul, token );
-		token = next_token( test_input, index );
-		EXPECT_EQ( 5ul, index );
-		EXPECT_EQ( 2ul, token );
-		token = next_token( test_input, index );
-		EXPECT_EQ( 7ul, index );
-		EXPECT_EQ( 5ul, token );
-		token = next_token( test_input, index );
-		EXPECT_EQ( 9ul, index );
-		EXPECT_EQ( 4ul, token );
-		token = next_token( test_input, index );
-		EXPECT_EQ( 10ul, index );
-		EXPECT_EQ( 0ul, token );
-		index = test_input.size();
-		token = next_token( test_input, index );
-		EXPECT_EQ( test_input.size() , index );
-		EXPECT_EQ( 1ul, token );
 
 	}
